Added solution to Problem 3110

+++ LeetcodeSolutions/Prob3110_ScoreOfAString.cpp
@@ -5,7 +5,7 @@ public:
         int score = 0;
         for(int i = 0;i<s.size()-1;i++)
         {
-            score+= abs((int)(s[i]-s[i+1]));
+            score+= abs(s[i]-s[i+1]);
         }
         return score;
     }
